remove myunit
