+++ mypy/test/testargs.py
@@ -5,8 +5,8 @@ defaults, and that argparse doesn't assign any new members to the Options
 object it creates.
 """
 
-from mypy.myunit import Suite, assert_equal
-from mypy.options import Options, BuildType
+from mypy.test.helpers import Suite
+from mypy.options import Options
 from mypy.main import process_options
 
 
@@ -14,4 +14,6 @@ class ArgSuite(Suite):
     def test_coherence(self) -> None:
         options = Options()
         _, parsed_options = process_options([], require_targets=False)
-        assert_equal(options, parsed_options)
+        assert options.__dict__.keys() == parsed_options.__dict__.keys()
+        for k in options.__dict__:
+            assert getattr(options, k) == getattr(parsed_options, k), k
