+++ mypy/test/testmerge.py
@@ -38,8 +38,8 @@ class ASTMergeSuite(DataSuite):
     base_path = test_temp_dir
     optional_out = True
 
-    def __init__(self, *, update_data: bool) -> None:
-        super().__init__(update_data=update_data)
+    def setup(self) -> None:
+        super().setup()
         self.str_conv = StrConv(show_ids=True)
         assert self.str_conv.id_mapper is not None
         self.id_mapper = self.str_conv.id_mapper  # type: IdMapper
