+++ mypy/test/testmoduleinfo.py
@@ -1,7 +1,5 @@
 from mypy import moduleinfo
-from mypy.myunit import (
-    Suite, assert_equal, assert_true, assert_false
-)
+from mypy.test.helpers import assert_true, assert_false, Suite
 
 
 class ModuleInfoSuite(Suite):
