+++ mypy/test/testreports.py
@@ -1,7 +1,7 @@
 """Test cases for reports generated by mypy."""
 import textwrap
 
-from mypy.myunit import Suite, assert_equal
+from mypy.test.helpers import Suite, assert_equal
 from mypy.report import CoberturaPackage, get_line_rate
 
 import lxml.etree as etree  # type: ignore
