+++ mypy/test/testsolve.py
@@ -2,7 +2,7 @@
 
 from typing import List, Union, Tuple, Optional
 
-from mypy.myunit import Suite, assert_equal
+from mypy.test.helpers import Suite, assert_equal
 from mypy.constraints import SUPERTYPE_OF, SUBTYPE_OF, Constraint
 from mypy.solve import solve_constraints
 from mypy.test.typefixture import TypeFixture
@@ -10,8 +10,7 @@ from mypy.types import Type, TypeVarType, TypeVarId
 
 
 class SolveSuite(Suite):
-    def __init__(self) -> None:
-        super().__init__()
+    def setUp(self) -> None:
         self.fx = TypeFixture()
 
     def test_empty_input(self) -> None:
