+++ pytest.ini
@@ -14,7 +14,7 @@ python_files = test*.py
 
 # Because we provide our own collection logic, disable the default
 # python collector by giving it empty patterns to search for.
-python_classes =
+python_classes = *Suite
 python_functions =
 
 # always run in parallel (requires pytest-xdist, see test-requirements.txt)
