+++ setup.py
@@ -100,7 +100,7 @@ setup(name='mypy',
       license='MIT License',
       platforms=['POSIX'],
       py_modules=[],
-      packages=['mypy', 'mypy.test', 'mypy.myunit', 'mypy.server'],
+      packages=['mypy', 'mypy.test', 'mypy.server'],
       entry_points={'console_scripts': ['mypy=mypy.__main__:console_entry',
                                         'stubgen=mypy.stubgen:main',
                                         'dmypy=mypy.dmypy:main',
